fix background issues, add comments to the code.
I hope you can understand it now. This is basicially a revert to the svn 599 code(or around).
The logic of bg-color and bg-pixmap:
If bg-pixmap != NULL bg-color == NULL, use bg-pixmap for bg
If bg-color != NULL bg-pixmap == NULL, use bg-color for bg
If bg-color == NULL bg-pixmap == NULL, use gtk theme for bg.

If bg-color != NULL && bg-pixmap != NULL, behavior is undefined. This is never going to happen.