Issue 203.
Fill with the pattern with REPEAT flag.

Although Gdk will tile the bg_pixmap,
Our cairo image cloner was filling in random bytes for the extended regions in the cloned pixmap, which always has the same geometry of the GdkWindow.
Because the geometry of the bg_pixmap is the same as the window, gdk won't think she needs to tile the bg, and would happily pick up the gabled bg_pixmap, painting an abstract art to the panel.