Tweak the gconfdialog. Hopefully now it strictly follows the construction procedure of of regualar GTK objects.
To Luca:

Vala CreateMethod is a magic. Saying if the object has property A, B

public A {get; set;}
public B {get; construct;}

and in the CreateMethod

public OBJ(int para) {
  A = para;
  B = para;
}
construct {
  message("%d", A);
  message("%d", B);
}

will show A = para, and B = 0(uninialized).
construct is executed after every 'construct' property is initialized and before every non-'construct' property is initialized.
Usually in GTK people initialized private members in _init. _init is invoked before 'construct' properties are initialized. but in Vala we don't have such luxury. the only place to inialize private members is construct block.

In GObject, g_object_new is the only legal way of creating objects. CreateMethod are by definition convinient wrappers: ie, if they are not there, the public members of that class should always be able to provide the same functionality.

If you want your library to work smoothly with applications written in raw C, this mess has to be cared.